+++ brendancuda_ai_evolution_evaluation_output_implementations_uniquevalues.h
@@ -14,7 +14,7 @@ namespace BrendanCUDA {
                         template <typename T>
                         struct evalUniqueValues_sd final {
                             instanceFunctions_t<T> instanceFunctions;
-                            size_t iterationCount;
+                            uint64_t iterationCount;
                             size_t outputCount;
                             bool individual;
                             void* sd_ci;
